projectRoot/requirements.txt
APScheduler==3.11.0
jsonpickle==4.0.1
langchain==0.3.18
langchain_anthropic==0.3.7
langchain_core==0.3.34
langchain_openai==0.3.4
#numpy==1.9.0
pandas==2.2.3
pymongo==4.11.1
python-dotenv==1.0.1
python-telegram-bot==21.10
regex==2024.5.15
scikit_learn==1.6.1
sentence_transformers==3.3.1
